move Battery, ElectricVehicle and ElectricFleet(Module) to fleet package
