+++ contribs/drt/src/main/java/org/matsim/contrib/edrt/optimizer/EDrtVehicleDataEntryFactory.java
@@ -32,7 +32,7 @@ import org.matsim.contrib.dvrp.schedule.Schedule.ScheduleStatus;
 import org.matsim.contrib.dvrp.schedule.Task;
 import org.matsim.contrib.dvrp.schedule.Task.TaskStatus;
 import org.matsim.contrib.edrt.schedule.EDrtTask;
-import org.matsim.contrib.ev.data.Battery;
+import org.matsim.contrib.ev.fleet.Battery;
 import org.matsim.contrib.ev.dvrp.ChargingTask;
 import org.matsim.contrib.ev.dvrp.EvDvrpVehicle;
 import org.matsim.contrib.ev.dvrp.tracker.ETaskTracker;
