+++ contribs/ev/src/main/java/org/matsim/contrib/ev/charging/ChargingEstimations.java
@@ -19,7 +19,7 @@
 package org.matsim.contrib.ev.charging;
 
 import org.matsim.contrib.ev.data.Charger;
-import org.matsim.contrib.ev.data.ElectricVehicle;
+import org.matsim.contrib.ev.fleet.ElectricVehicle;
 
 import java.util.stream.Stream;
