+++ contribs/ev/src/main/java/org/matsim/contrib/ev/charging/ChargingListener.java
@@ -18,7 +18,7 @@
 
 package org.matsim.contrib.ev.charging;
 
-import org.matsim.contrib.ev.data.ElectricVehicle;
+import org.matsim.contrib.ev.fleet.ElectricVehicle;
 
 /**
  * @author michalm
