+++ contribs/ev/src/main/java/org/matsim/contrib/ev/charging/ChargingLogic.java
@@ -20,7 +20,7 @@
 package org.matsim.contrib.ev.charging;
 
 import org.matsim.contrib.ev.data.Charger;
-import org.matsim.contrib.ev.data.ElectricVehicle;
+import org.matsim.contrib.ev.fleet.ElectricVehicle;
 import org.matsim.core.api.experimental.events.EventsManager;
 
 import java.util.Collection;
