+++ contribs/ev/src/main/java/org/matsim/contrib/ev/charging/ChargingWithQueueingAndAssignmentLogic.java
@@ -21,7 +21,7 @@ package org.matsim.contrib.ev.charging;
 
 import org.matsim.api.core.v01.Id;
 import org.matsim.contrib.ev.data.Charger;
-import org.matsim.contrib.ev.data.ElectricVehicle;
+import org.matsim.contrib.ev.fleet.ElectricVehicle;
 
 import java.util.Collection;
 import java.util.Collections;
