+++ contribs/ev/src/main/java/org/matsim/contrib/ev/charging/FixedSpeedChargingStrategy.java
@@ -18,8 +18,8 @@
 
 package org.matsim.contrib.ev.charging;
 
-import org.matsim.contrib.ev.data.Battery;
-import org.matsim.contrib.ev.data.ElectricVehicle;
+import org.matsim.contrib.ev.fleet.Battery;
+import org.matsim.contrib.ev.fleet.ElectricVehicle;
 
 /**
  * @author michalm
