+++ contribs/ev/src/main/java/org/matsim/contrib/ev/dvrp/ChargingActivity.java
@@ -21,7 +21,7 @@ package org.matsim.contrib.ev.dvrp;
 
 import org.matsim.contrib.dynagent.DynActivity;
 import org.matsim.contrib.ev.charging.ChargingWithQueueingAndAssignmentLogic;
-import org.matsim.contrib.ev.data.ElectricVehicle;
+import org.matsim.contrib.ev.fleet.ElectricVehicle;
 
 public class ChargingActivity implements DynActivity {
 	private static final String CHARGING_ACTIVITY_TYPE = "ChargingActivity";
