+++ contribs/ev/src/main/java/org/matsim/contrib/ev/dvrp/ChargingTaskImpl.java
@@ -21,7 +21,7 @@ package org.matsim.contrib.ev.dvrp;
 import org.matsim.contrib.dvrp.schedule.StayTaskImpl;
 import org.matsim.contrib.ev.charging.ChargingWithQueueingAndAssignmentLogic;
 import org.matsim.contrib.ev.data.Charger;
-import org.matsim.contrib.ev.data.ElectricVehicle;
+import org.matsim.contrib.ev.fleet.ElectricVehicle;
 
 /**
  * @author michalm
