+++ contribs/ev/src/main/java/org/matsim/contrib/ev/dvrp/DvrpChargingListener.java
@@ -19,7 +19,7 @@
 package org.matsim.contrib.ev.dvrp;
 
 import org.matsim.contrib.ev.charging.ChargingListener;
-import org.matsim.contrib.ev.data.ElectricVehicle;
+import org.matsim.contrib.ev.fleet.ElectricVehicle;
 
 /**
  * @author michalm
