+++ contribs/ev/src/main/java/org/matsim/contrib/ev/fleet/ElectricFleet.java
@@ -1,9 +1,9 @@
-/* *********************************************************************** *
+/*
+ * *********************************************************************** *
  * project: org.matsim.*
- *                                                                         *
  * *********************************************************************** *
  *                                                                         *
- * copyright       : (C) 2013 by the members listed in the COPYING,        *
+ * copyright       : (C) 2019 by the members listed in the COPYING,        *
  *                   LICENSE and WARRANTY file.                            *
  * email           : info at matsim dot org                                *
  *                                                                         *
@@ -15,28 +15,20 @@
  *   (at your option) any later version.                                   *
  *   See also COPYING, LICENSE and WARRANTY file                           *
  *                                                                         *
- * *********************************************************************** */
+ * *********************************************************************** *
+ */
 
-package org.matsim.contrib.ev.data;
+package org.matsim.contrib.ev.fleet;
 
-import java.util.List;
-
-import org.matsim.api.core.v01.Identifiable;
+import org.matsim.api.core.v01.Id;
 import org.matsim.contrib.ev.discharging.AuxEnergyConsumption;
 import org.matsim.contrib.ev.discharging.DriveEnergyConsumption;
 
-public interface ElectricVehicle extends Identifiable<ElectricVehicle> {
-	DriveEnergyConsumption getDriveEnergyConsumption();
+import java.util.Map;
 
-	AuxEnergyConsumption getAuxEnergyConsumption();
+public interface ElectricFleet {
+	Map<Id<ElectricVehicle>, ElectricVehicle> getElectricVehicles();
 
-	Battery getBattery();
-
-	List<String> getChargerTypes();
-
-	String getVehicleType();
-
-	void setDriveEnergyConsumption(DriveEnergyConsumption driveEnergyConsumption);
-
-	void setAuxEnergyConsumption(AuxEnergyConsumption auxEnergyConsumption);
+	void resetBatteriesAndConsumptions(DriveEnergyConsumption.Factory driveConsumptionFactory,
+			AuxEnergyConsumption.Factory auxConsumptionFactory);
 }
