+++ contribs/ev/src/main/java/org/matsim/contrib/ev/fleet/ElectricFleetReader.java
@@ -26,8 +26,6 @@ import java.util.Stack;
 import org.matsim.api.core.v01.Id;
 import org.matsim.contrib.ev.EvUnits;
 import org.matsim.contrib.ev.data.ChargerImpl;
-import org.matsim.contrib.ev.data.ElectricVehicle;
-import org.matsim.contrib.ev.data.ElectricVehicleImpl;
 import org.matsim.core.utils.io.MatsimXmlParser;
 import org.xml.sax.Attributes;
