+++ contribs/ev/src/main/java/org/matsim/contrib/ev/routing/EVNetworkRoutingModule.java
@@ -35,8 +35,8 @@ import org.matsim.api.core.v01.population.PlanElement;
 import org.matsim.contrib.ev.charging.VehicleChargingHandler;
 import org.matsim.contrib.ev.data.Charger;
 import org.matsim.contrib.ev.data.ChargingInfrastructure;
-import org.matsim.contrib.ev.data.ElectricFleet;
-import org.matsim.contrib.ev.data.ElectricVehicle;
+import org.matsim.contrib.ev.fleet.ElectricFleet;
+import org.matsim.contrib.ev.fleet.ElectricVehicle;
 import org.matsim.contrib.util.LinkProvider;
 import org.matsim.contrib.util.StraightLineKnnFinder;
 import org.matsim.core.gbl.Gbl;
