+++ contribs/ev/src/main/java/org/matsim/contrib/ev/stats/IndividualSocTimeProfileCollectorProvider.java
@@ -22,8 +22,8 @@ package org.matsim.contrib.ev.stats;
 import com.google.inject.Inject;
 import com.google.inject.Provider;
 import org.matsim.contrib.ev.EvUnits;
-import org.matsim.contrib.ev.data.ElectricFleet;
-import org.matsim.contrib.ev.data.ElectricVehicle;
+import org.matsim.contrib.ev.fleet.ElectricFleet;
+import org.matsim.contrib.ev.fleet.ElectricVehicle;
 import org.matsim.contrib.util.timeprofile.TimeProfileCollector;
 import org.matsim.contrib.util.timeprofile.TimeProfileCollector.ProfileCalculator;
 import org.matsim.contrib.util.timeprofile.TimeProfiles;
