+++ contribs/ev/src/main/java/org/matsim/contrib/ev/stats/VehicleTypeAggregatedSocTimeProfileCollectorProvider.java
@@ -22,7 +22,7 @@ package org.matsim.contrib.ev.stats;
 import com.google.inject.Inject;
 import com.google.inject.Provider;
 import org.matsim.contrib.ev.EvUnits;
-import org.matsim.contrib.ev.data.ElectricFleet;
+import org.matsim.contrib.ev.fleet.ElectricFleet;
 import org.matsim.contrib.util.timeprofile.TimeProfileCollector;
 import org.matsim.contrib.util.timeprofile.TimeProfileCollector.ProfileCalculator;
 import org.matsim.contrib.util.timeprofile.TimeProfiles;
